Don't destroy export menu when changing sidebar selection
Destroying the menu causes a crash under Linux due to the menu event from the menu lingering for some reason. Now, construct the menu once and just update the label on the "Export..." menu item.
This also refactor a lot of redundant code and removes menu logic unrelated to the Edit bar from the Edit bar section.